+++ app/controllers/premiumlease/CalculatedFigureYourselfController.scala
@@ -14,40 +14,40 @@
  * limitations under the License.
  */
 
-package controllers.premiumLease
+package controllers.premiumlease
 
 import controllers.actions._
-import forms.premiumLease.PremiumsGrantLeaseFormProvider
+import forms.premiumlease.CalculatedFigureYourselfFormProvider
 import models.Mode
 import navigation.Navigator
-import pages.premiumLease.PremiumsGrantLeasePage
+import pages.CalculatedFigureYourselfPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.premiumLease.PremiumsGrantLeaseView
+import views.html.premiumlease.CalculatedFigureYourselfView
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class PremiumsGrantLeaseController @Inject()(
-                                        override val messagesApi: MessagesApi,
-                                        sessionRepository: SessionRepository,
-                                        navigator: Navigator,
-                                        identify: IdentifierAction,
-                                        getData: DataRetrievalAction,
-                                        requireData: DataRequiredAction,
-                                        formProvider: PremiumsGrantLeaseFormProvider,
-                                        val controllerComponents: MessagesControllerComponents,
-                                        view: PremiumsGrantLeaseView
-                                      )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
+class CalculatedFigureYourselfController @Inject()(
+                                         override val messagesApi: MessagesApi,
+                                         sessionRepository: SessionRepository,
+                                         navigator: Navigator,
+                                         identify: IdentifierAction,
+                                         getData: DataRetrievalAction,
+                                         requireData: DataRequiredAction,
+                                         formProvider: CalculatedFigureYourselfFormProvider,
+                                         val controllerComponents: MessagesControllerComponents,
+                                         view: CalculatedFigureYourselfView
+                                 )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
 
   val form = formProvider()
 
   def onPageLoad(taxYear: Int, mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) {
     implicit request =>
 
-      val preparedForm = request.userAnswers.get(PremiumsGrantLeasePage) match {
+      val preparedForm = request.userAnswers.get(CalculatedFigureYourselfPage) match {
         case None => form
         case Some(value) => form.fill(value)
       }
@@ -60,13 +60,13 @@ class PremiumsGrantLeaseController @Inject()(
 
       form.bindFromRequest().fold(
         formWithErrors =>
-          Future.successful(BadRequest(view(formWithErrors,taxYear, mode))),
+          Future.successful(BadRequest(view(formWithErrors, taxYear, mode))),
 
         value =>
           for {
-            updatedAnswers <- Future.fromTry(request.userAnswers.set(PremiumsGrantLeasePage, value))
+            updatedAnswers <- Future.fromTry(request.userAnswers.set(CalculatedFigureYourselfPage, value))
             _              <- sessionRepository.set(updatedAnswers)
-          } yield Redirect(navigator.nextPage(PremiumsGrantLeasePage, taxYear ,mode, updatedAnswers))
+          } yield Redirect(navigator.nextPage(CalculatedFigureYourselfPage, taxYear, mode, updatedAnswers))
       )
   }
 }
