+++ app/forms/premiumlease/PremiumsGrantLeaseFormProvider.scala
@@ -14,17 +14,21 @@
  * limitations under the License.
  */
 
-package forms.premiumLease
+package forms.premiumlease
 
 import forms.mappings.Mappings
 import play.api.data.Form
 
 import javax.inject.Inject
 
-class LeasePremiumPaymentFormProvider @Inject() extends Mappings {
+class PremiumsGrantLeaseFormProvider @Inject() extends Mappings {
 
-  def apply(): Form[Boolean] =
+  def apply(): Form[Int] =
     Form(
-      "value" -> boolean("leasePremiumPayment.error.required")
+      "value" -> int(
+        "premiumsGrantLease.error.required",
+        "premiumsGrantLease.error.wholeNumber",
+        "premiumsGrantLease.error.nonNumeric")
+          .verifying(inRange(0, 1000000000, "premiumsGrantLease.error.outOfRange"))
     )
 }
