+++ app/viewmodels/checkAnswers/premiumlease/CalculatedFigureYourselfSummary.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package viewmodels.checkAnswers.premiumLease
+package viewmodels.checkAnswers.premiumlease
 
 import models.{CheckMode, UserAnswers}
 import pages.CalculatedFigureYourselfPage
@@ -22,7 +22,7 @@ import play.api.i18n.Messages
 import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist.SummaryListRow
 import viewmodels.govuk.summarylist._
 import viewmodels.implicits._
-import controllers.premiumLease._
+import controllers.premiumlease._
 object CalculatedFigureYourselfSummary {
 
   def row(taxYear: Int, answers: UserAnswers)(implicit messages: Messages): Option[SummaryListRow] =
