+++ app/views/premiumlease/CalculatedFigureYourselfView.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import controllers.premiumLease.routes._
+@import controllers.premiumlease.routes._
 
 @this(
     layout: templates.Layout,
