+++ app/views/premiumlease/LeasePremiumPaymentView.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import viewmodels.LegendSize
-@import controllers.premiumLease.routes._
+@import controllers.premiumlease.routes._
 
 @this(
     layout: templates.Layout,
