+++ app/views/premiumlease/PremiumsGrantLeaseView.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import viewmodels.InputWidth._
-@import controllers.premiumLease.routes._
+@import controllers.premiumlease.routes._
 
 
 @this(
