+++ test/controllers/premiumlease/CalculatedFigureYourselfControllerSpec.scala
@@ -1,48 +1,61 @@
-package controllers.premiumLease
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers.premiumlease
 
 import base.SpecBase
-import forms.premiumLease.PremiumsGrantLeaseFormProvider
+import forms.premiumlease.CalculatedFigureYourselfFormProvider
 import models.{NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.premiumLease.PremiumsGrantLeasePage
+import pages.CalculatedFigureYourselfPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.premiumLease.PremiumsGrantLeaseView
+import controllers.premiumlease.CalculatedFigureYourselfView
 
 import java.time.LocalDate
 import scala.concurrent.Future
 
-class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
-
-  val formProvider = new PremiumsGrantLeaseFormProvider()
-  val form = formProvider()
-  private val taxYear = LocalDate.now.getYear
-
+class CalculatedFigureYourselfControllerSpec extends SpecBase with MockitoSugar {
 
   def onwardRoute = Call("GET", "/foo")
 
-  val validAnswer = 0
+  val formProvider = new CalculatedFigureYourselfFormProvider()
+  val form = formProvider()
+  val taxYear = LocalDate.now.getYear
 
-  lazy val premiumsGrantLeaseRoute = routes.PremiumsGrantLeaseController.onPageLoad(taxYear, NormalMode).url
+  lazy val calculatedFigureYourselfRoute = routes.CalculatedFigureYourselfController.onPageLoad(taxYear, NormalMode).url
 
-  "PremiumsGrantLease Controller" - {
+  "CalculatedFigureYourself Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
       val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, taxYear, NormalMode)(request, messages(application)).toString
@@ -51,19 +64,19 @@ class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val userAnswers = UserAnswers(userAnswersId).set(PremiumsGrantLeasePage, validAnswer).success.value
+      val userAnswers = UserAnswers(userAnswersId).set(CalculatedFigureYourselfPage, true).success.value
 
       val application = applicationBuilder(userAnswers = Some(userAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
-        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         val result = route(application, request).value
 
         status(result) mustEqual OK
-        contentAsString(result) mustEqual view(form.fill(validAnswer), taxYear, NormalMode)(request, messages(application)).toString
+        contentAsString(result) mustEqual view(form.fill(true), taxYear, NormalMode)(request, messages(application)).toString
       }
     }
 
@@ -83,8 +96,8 @@ class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, premiumsGrantLeaseRoute)
-            .withFormUrlEncodedBody(("value", validAnswer.toString))
+          FakeRequest(POST, calculatedFigureYourselfRoute)
+            .withFormUrlEncodedBody(("value", "true"))
 
         val result = route(application, request).value
 
@@ -99,12 +112,12 @@ class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, premiumsGrantLeaseRoute)
-            .withFormUrlEncodedBody(("value", "invalid value"))
+          FakeRequest(POST, calculatedFigureYourselfRoute)
+            .withFormUrlEncodedBody(("value", ""))
 
-        val boundForm = form.bind(Map("value" -> "invalid value"))
+        val boundForm = form.bind(Map("value" -> ""))
 
-        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
+        val view = application.injector.instanceOf[CalculatedFigureYourselfView]
 
         val result = route(application, request).value
 
@@ -118,7 +131,7 @@ class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
       val application = applicationBuilder(userAnswers = None, true).build()
 
       running(application) {
-        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
+        val request = FakeRequest(GET, calculatedFigureYourselfRoute)
 
         val result = route(application, request).value
 
@@ -133,13 +146,12 @@ class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, premiumsGrantLeaseRoute)
-            .withFormUrlEncodedBody(("value", validAnswer.toString))
+          FakeRequest(POST, calculatedFigureYourselfRoute)
+            .withFormUrlEncodedBody(("value", "true"))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-
         redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
