+++ test/controllers/premiumlease/PremiumsGrantLeaseControllerSpec.scala
@@ -14,50 +14,51 @@
  * limitations under the License.
  */
 
-package controllers.premiumLease
+package controllers.premiumlease
 
 import base.SpecBase
-import controllers.routes
-import forms.premiumLease.LeasePremiumPaymentFormProvider
+import forms.premiumlease.PremiumsGrantLeaseFormProvider
 import models.{NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.premiumLease.LeasePremiumPaymentPage
+import pages.premiumLease.PremiumsGrantLeasePage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.premiumLease.LeasePremiumPaymentView
-import controllers.premiumLease.routes._
+import controllers.premiumlease.PremiumsGrantLeaseView
 
 import java.time.LocalDate
 import scala.concurrent.Future
 
-class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
+class PremiumsGrantLeaseControllerSpec extends SpecBase with MockitoSugar {
 
-  def onwardRoute : Call = Call("GET", "/foo")
-
-  val formProvider = new LeasePremiumPaymentFormProvider()
+  val formProvider = new PremiumsGrantLeaseFormProvider()
   val form = formProvider()
   private val taxYear = LocalDate.now.getYear
 
-  lazy val leasePremiumPaymentRoute = LeasePremiumPaymentController.onPageLoad(taxYear, NormalMode).url
 
-  "LeasePremiumPayment Controller" - {
+  def onwardRoute = Call("GET", "/foo")
+
+  val validAnswer = 0
+
+  lazy val premiumsGrantLeaseRoute = routes.PremiumsGrantLeaseController.onPageLoad(taxYear, NormalMode).url
+
+  "PremiumsGrantLease Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
-      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, taxYear, NormalMode)(request, messages(application)).toString
@@ -66,19 +67,19 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val userAnswers = UserAnswers(userAnswersId).set(LeasePremiumPaymentPage, true).success.value
+      val userAnswers = UserAnswers(userAnswersId).set(PremiumsGrantLeasePage, validAnswer).success.value
 
-      val application = applicationBuilder(userAnswers = Some(userAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(userAnswers), false).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
 
         val result = route(application, request).value
 
         status(result) mustEqual OK
-        contentAsString(result) mustEqual view(form.fill(true), taxYear, NormalMode)(request, messages(application)).toString
+        contentAsString(result) mustEqual view(form.fill(validAnswer), taxYear, NormalMode)(request, messages(application)).toString
       }
     }
 
@@ -89,7 +90,7 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
       when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
       val application =
-        applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false)
+        applicationBuilder(userAnswers = Some(emptyUserAnswers), false)
           .overrides(
             bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
             bind[SessionRepository].toInstance(mockSessionRepository)
@@ -98,8 +99,8 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
-            .withFormUrlEncodedBody(("value", "true"))
+          FakeRequest(POST, premiumsGrantLeaseRoute)
+            .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
 
@@ -110,16 +111,16 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must return a Bad Request and errors when invalid data is submitted" in {
 
-      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), isAgent = false).build()
+      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers), false).build()
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
-            .withFormUrlEncodedBody(("value", ""))
+          FakeRequest(POST, premiumsGrantLeaseRoute)
+            .withFormUrlEncodedBody(("value", "invalid value"))
 
-        val boundForm = form.bind(Map("value" -> ""))
+        val boundForm = form.bind(Map("value" -> "invalid value"))
 
-        val view = application.injector.instanceOf[LeasePremiumPaymentView]
+        val view = application.injector.instanceOf[PremiumsGrantLeaseView]
 
         val result = route(application, request).value
 
@@ -130,31 +131,32 @@ class LeasePremiumPaymentControllerSpec extends SpecBase with MockitoSugar {
 
     "must redirect to Journey Recovery for a GET if no existing data is found" in {
 
-      val application = applicationBuilder(userAnswers = None, isAgent = true).build()
+      val application = applicationBuilder(userAnswers = None, true).build()
 
       running(application) {
-        val request = FakeRequest(GET, leasePremiumPaymentRoute)
+        val request = FakeRequest(GET, premiumsGrantLeaseRoute)
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
 
     "must redirect to Journey Recovery for a POST if no existing data is found" in {
 
-      val application = applicationBuilder(userAnswers = None, isAgent = true).build()
+      val application = applicationBuilder(userAnswers = None, true).build()
 
       running(application) {
         val request =
-          FakeRequest(POST, leasePremiumPaymentRoute)
-            .withFormUrlEncodedBody(("value", "true"))
+          FakeRequest(POST, premiumsGrantLeaseRoute)
+            .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
   }
